One change

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales= {10, 20, 30};
+        long[] sales = {10, 20, 30};
         SalesManager manager = new SalesManager(sales);
         System.out.println(manager.max());
     }
